+++ build.clj
@@ -4,7 +4,7 @@
    [deps-deploy.deps-deploy :as dd]))
 
 (def lib 'com.adgoji/mollie)
-(def version "0.1.4")
+(def version "0.2.0")
 (def class-dir "target/classes")
 (def basis (b/create-basis {:project "deps.edn"}))
 (def jar-file (format "target/%s-%s.jar" (name lib) version))