+++ src/com/adgoji/mollie/api/mandates.clj
@@ -1,14 +1,13 @@
 (ns com.adgoji.mollie.api.mandates
   (:require
-   [clojure.spec.alpha :as s]
-   [com.adgoji.mollie.client :as mollie.client]
    [com.adgoji.mollie :as mollie]
+   [com.adgoji.mollie.client :as mollie.client]
    [com.adgoji.mollie.creditcard :as creditcard]
    [com.adgoji.mollie.customer :as customer]
    [com.adgoji.mollie.directdebit :as directdebit]
    [com.adgoji.mollie.link :as link]
    [com.adgoji.mollie.mandate :as mandate]
-   [com.adgoji.mollie.mandate.request :as request]
+   [com.adgoji.mollie.mandate.request :as mandate.request]
    [com.adgoji.mollie.pagination :as pagination]
    [com.adgoji.mollie.paypal :as paypal]
    [com.adgoji.utils.spec :as spec])
@@ -51,46 +50,42 @@
              ::mandate/mode              (keyword mode)
              ::mandate/status            (keyword status)
              ::mandate/method            (keyword method)
-             ::mandate/details           details
              ::mandate/mandate-reference mandate-reference
              ::mandate/created-at        (Instant/parse created-at)
              ::link/self                 (spec/qualify-link (:self links))
              ::link/customer             (spec/qualify-link (:customer links))
              ::link/documentation        (spec/qualify-link (:documentation links))}
-      signature-date (assoc ::mandate/signature-date (LocalDate/parse signature-date)))))
+      signature-date (assoc ::mandate/signature-date (LocalDate/parse signature-date))
+      :always        (into details))))
 
 (defn create
   "Create a new mandate for given `customer-id`."
   [client customer-id mandate]
-  {:pre [(s/valid? ::customer/id customer-id)
-         (s/valid? ::request/create mandate)]}
-  (mollie.client/http-post client
-                           (format "/v2/customers/%s/mandates" customer-id)
-                           {:body                 mandate
-                            :response-transformer transform-mandate
-                            :spec                 ::mollie/mandate}))
+  (let [body (spec/check mandate ::mandate.request/create)]
+    (mollie.client/http-post client
+                             (format "/v2/customers/%s/mandates"
+                                     (spec/check customer-id ::customer/id))
+                             {:body                 body
+                              :response-transformer transform-mandate
+                              :spec                 ::mollie/mandate})))
 
 (defn get-by-id
   "Fetch a single customer's mandate by `mandate-id`."
   [client customer-id mandate-id]
-  {:pre [(s/valid? ::customer/id customer-id)
-         (s/valid? ::mandate/id mandate-id)]}
   (mollie.client/http-get client
                           (format "/v2/customers/%s/mandates/%s"
-                                  customer-id
-                                  mandate-id)
+                                  (spec/check customer-id ::customer/id)
+                                  (spec/check mandate-id ::mandate/id))
                           {:response-transformer transform-mandate
                            :spec                 ::mollie/mandate}))
 
 (defn revoke-by-id
   "Revoke a specific customer's mandate by `mandate-id`."
   [client customer-id mandate-id]
-  {:pre [(s/valid? ::customer/id customer-id)
-         (s/valid? ::mandate/id mandate-id)]}
   (mollie.client/http-delete client
                              (format "/v2/customers/%s/mandates/%s"
-                                     customer-id
-                                     mandate-id)))
+                                     (spec/check customer-id ::customer/id)
+                                     (spec/check mandate-id ::mandate/id))))
 
 (defn- transform-mandates
   [response]
@@ -110,14 +105,17 @@
 
 (defn get-list
   "Fetch all customer's mandates."
-  [client customer-id {:keys [from limit] :as opts}]
-  {:pre [(s/valid? ::customer/id customer-id)
-         (s/valid? (s/keys :opt-un [::pagination/from ::pagination/limit]) opts)]}
-  (let [fetch-fn (if limit
-                   mollie.client/http-get
-                   (partial mollie.client/fetch-all ::mollie/mandates))]
+  [client customer-id opts]
+  (let [{:keys [from limit]}
+        (spec/check opts ::pagination/opts)
+
+        fetch-fn
+        (if limit
+          mollie.client/http-get
+          (partial mollie.client/fetch-all ::mollie/mandates))]
     (fetch-fn client
-              (format "/v2/customers/%s/mandates" customer-id)
+              (format "/v2/customers/%s/mandates"
+                      (spec/check customer-id ::customer/id))
               {:response-transformer transform-mandates
                :spec                 ::mollie/mandates-list
                :query-params         (cond-> {}